requirements.txt
sqlalchemy>=0.9, <0.9.99
dateutils>=0.6.6
icalendar==3.6.1
python-dateutil==2.2
pycrypto==2.6.1
setproctitle==1.1.8
coverage
